requirements.txt
stripe
pinecone
langchain
langchain-community
langchain-core
langgraph
langchain-anthropic
langsmith
langchain-aws
langchain-pinecone
pinecone-notebooks
langchain-xai
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
daphne
channels
psycopg2-binary
dj-database-url
aenum==3.1.15 ; python_version >= "3.10" and python_version < "4.0"
aiofiles==24.1.0 ; python_version >= "3.10" and python_version < "4.0"
aiohappyeyeballs==2.3.5 ; python_version >= "3.10" and python_version < "4.0"
aiohttp==3.10.1 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anthropic==0.19.2 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.11"
attrs==24.2.0 ; python_version >= "3.10" and python_version < "4.0"
autobahn==24.4.2 ; python_version >= "3.10" and python_version < "4.0"
automat==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.0 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
channels==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
constantly==23.10.4 ; python_version >= "3.10" and python_version < "4.0"
cryptography==43.0.0 ; python_version >= "3.10" and python_version < "4.0"
daphne==4.1.2 ; python_version >= "3.10" and python_version < "4.0"
dataclasses-json==0.6.7 ; python_version >= "3.10" and python_version < "4.0"
deepgram-sdk==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
dj-database-url==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
django==5.1 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-simplejwt==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.15.4 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2024.6.1 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
huggingface-hub==0.24.5 ; python_version >= "3.10" and python_version < "4.0"
hyperlink==21.0.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.7 ; python_version >= "3.10" and python_version < "4.0"
incremental==24.7.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
marshmallow==3.21.3 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
px-settings==0.1.5 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.20.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
pyopenssl==24.2.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
service-identity==24.1.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==72.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
tokenizers==0.19.1 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
tqdm==4.66.5 ; python_version >= "3.10" and python_version < "4.0"
twisted-iocpsupport==1.0.4 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
twisted[tls]==24.3.0 ; python_version >= "3.10" and python_version < "4.0"
txaio==23.1.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
wc-html2pdf==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.10" and python_version < "4.0"
zope-interface==7.0.1 ; python_version >= "3.10" and python_version < "4.0"
